Improve filter handling
1. Only try to process filters if there are some to process
2. Add startProcessingAdhellFilters to its own try/catch block
3. Add public / private methods in BlockUrlPatternsMatch.java to fit with the rest of the code
4. Don't try to add or remove domains during the filter processing if there aren't any

+++ app/src/main/java/com/fusionjack/adhell3/utils/BlockUrlUtils.java
@@ -80,7 +80,7 @@ public class BlockUrlUtils {
         // For each filter
         for(BlockUrl filter : filterSyntaxs){
             // Match filter syntax (so we can extract necessary info)
-            final Matcher filterMatcher = BlockUrlPatternsMatch.matchFilterSyntax(filter.url);
+            final Matcher filterMatcher = BlockUrlPatternsMatch.getmatchFilterSyntax(filter.url);
             // If there were matches (is a valid filter domain)
             if(filterMatcher.matches()){
                 final String delimiter = filterMatcher.group(1);
@@ -97,11 +97,16 @@ public class BlockUrlUtils {
                 }
             }
         }
-        // Add necessary domains
-        appDatabase.blockUrlDao().insertAll(blockUrls);
-        // Remove necessary domains
-        for(String removal : filterRemovals){
-            appDatabase.blockUrlDao().deleteBlockUrl(removal);
+        // Conditionally add blockurls
+        if(!blockUrls.isEmpty()){
+            appDatabase.blockUrlDao().insertAll(blockUrls);
+        }
+
+        // Conditionally remove blockurls
+        if(!filterRemovals.isEmpty()) {
+            for (String removal : filterRemovals) {
+                appDatabase.blockUrlDao().deleteBlockUrl(removal);
+            }
         }
     }
